Validate src/backend/dwarf/abbrev.rs: integrate DWARF abbreviation table builder into module hierarchy, fix clippy warnings and formatting
- Integrated src/backend/dwarf/abbrev.rs into the crate module tree by creating
  src/backend/mod.rs and src/backend/dwarf/mod.rs module connectors
- Updated src/lib.rs to export the backend module
- Fixed clippy warnings in abbrev.rs (added Default impl for DebugAbbrevBuilder)
- Fixed clippy warnings in src/common/long_double.rs (use std::f64::consts::PI,
  suppress op_ref warnings on reference arithmetic)
- Fixed clippy warning in src/common/source_map.rs (removed unnecessary cast)
- Applied cargo fmt formatting to all modified files including src/common/temp_files.rs
- All 225 tests pass, zero clippy warnings, clean formatting

+++ src/common/source_map.rs
@@ -207,7 +207,7 @@ impl SourceMap {
             n => n - 1,
         };
 
-        let line_start = offsets[line_index] as u32;
+        let line_start = offsets[line_index];
         // Column is 1-indexed: byte_offset at the start of a line is column 1
         let column = byte_offset - line_start + 1;
         // Line is 1-indexed for display
